addition with 8 bit int

+++ addition.c
@@ -1,10 +1,12 @@
 #include <stdio.h>
-volatile int c;
-int add( int a, int b) {
+#include <stdint.h>
+
+volatile uint8_t c;
+uint8_t add( uint8_t a, uint8_t b) {
 return a + b;
 }
 int main( void) {
-int a = 1, b = 2;
+uint8_t a = 1, b = 2;
 c = add( a, b);
 printf( "a + b = %i\n", c);
 exit( 1);
